+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 8.0.0 (upcoming)
+
+* Pending changelog
+
 ## 7.0.0 (upcoming)
 
 **Fixed bugs:**